pyproject.toml
[project]
name = "calculator-agent"
version = "0.1.0"
description = "Calculator agent using Ollama + Pydantic, no frameworks"
requires-python = ">=3.12"

dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.32.0",
    "ollama>=0.4.0",
    "bs4",
    "googlesearch-python>=1.2.0"
    
    ]